requirements.txt
streamlit>=1.36.0
openai>=1.40.0
requests>=2.31.0
python-docx>=1.1.2
pdfplumber>=0.11.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
